change: DynamoDBの型に合わせるために集約のバージョンの型をu64に変更する

+++ services/order/internal/kernel/src/command/model/aggregate.rs
@@ -7,12 +7,12 @@ use crate::id::Id;
 pub struct Aggregate {
     id: Id<Aggregate>,
     /// 集約のバージョン
-    version: u128,
+    version: u64,
 }
 
 impl Aggregate {
     #[must_use]
-    pub fn new(id: Id<Aggregate>, version: u128) -> Self {
+    pub fn new(id: Id<Aggregate>, version: u64) -> Self {
         Self { id, version }
     }
 
@@ -24,7 +24,7 @@ impl Aggregate {
 
     /// 集約のバージョン
     #[must_use]
-    pub fn version(&self) -> u128 {
+    pub fn version(&self) -> u64 {
         self.version
     }
